requirements.txt
beautifulsoup4==4.12.3
cloudscraper==1.2.71
python-dotenv==1.0.1
